Simplify Demo application
